requirements.txt
aiodns
aiohttp
censys
asyncio
beautifulsoup4
dnspython
logging
requests
requests_html